+++ CsharpEmployee/Employee/IEmployeeRepository.cs
@@ -14,7 +14,7 @@ namespace CsharpEmployee.Employee
         Task<PagedResult<EmployeeEntity>> GetEmployeesPagedFilteredAsync(
         int pageNumber, int pageSize, string term, bool archived);        Task<EmployeeEntity> GetByIdAsync(long id);
         Task<EmployeeEntity> UpdateEmployee(EmployeeEntity employee);
-        Task DeleteEmployee(int id);
+        Task DeleteEmployee(long id);
         bool ExistsByEmployeeUser(string newUser);
     }
 }
